+++ src/OpenTelemetry.Resources.Process/CHANGELOG.md
@@ -2,11 +2,20 @@
 
 ## Unreleased
 
+## 1.16.0-rc.1
+
+Released 2026-Jul-09
+
 * Updated OpenTelemetry core component version(s) to `1.16.0`.
   ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
 
 * Assemblies are now digitally signed using cosign.
   ([#4637](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4637))
+
+* Updated semantic conventions to
+  [v1.43.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.43.0/docs/resource/process.md)
+  and marked package as release candidate.
+  ([#4675](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4675))
 
 ## 1.15.1-beta.2